requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain-core==0.1.22
langchain-mistralai==0.0.3
pydantic==2.5.3 